refactor: move filterFavorites function to utils
