+++ src/getFavourites/index.ts
@@ -1,18 +1,13 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import * as Redis from 'ioredis';
 import { JSONSchemaType } from 'ajv';
-import {
-  RedisSettings,
-  Cache,
-  Favourite,
-  Favourites,
-  GetSchema,
-} from '../util/types';
+import { RedisSettings, Cache, GetSchema } from '../util/types';
 import { AzureFunction, Context, HttpRequest } from '@azure/functions';
 import validate from '../util/validator';
 import createErrorResponse from '../util/createErrorResponse';
 import { getDataStorage, getFavorites } from '../agent/Agent';
 import { getRedisHost, getRedisPort, getRedisPass } from '../util/helpers';
+import filterFavorites from '../util/filterFavorites';
 
 const getSchema: JSONSchemaType<GetSchema> = {
   type: 'object',
@@ -27,27 +22,6 @@ const getSchema: JSONSchemaType<GetSchema> = {
   required: ['hslId', 'store'],
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
 } as any;
-
-const filterFavorites = (
-  favorites: Favourites,
-  type: string | undefined = undefined,
-): Array<Favourite> => {
-  const keys = Object.keys(favorites);
-  const responseArray: Array<Favourite> = keys.map((key: string) => {
-    return Object(favorites)[key];
-  });
-  const filteredArray: Array<Favourite> = responseArray.filter(item => {
-    const itemType = String(item.type);
-    if (
-      item &&
-      ((!type && itemType !== 'note') || (type && itemType === type))
-    ) {
-      return true;
-    }
-    return false;
-  });
-  return filteredArray;
-};
 
 const getFavoritesTrigger: AzureFunction = async function (
   context: Context,
